+++ src/interpreter/type_check.rs
@@ -303,7 +303,7 @@ pub(super) fn type_implements_trait(
         "IntoIterator" if matches!(actual, Type::Named { name, arguments } if name == "Vec" && arguments.len() == 1) => {
             true
         }
-        "Iterator" if matches!(actual, Type::Named { name, arguments } if name == "SequenceIterator" && arguments.len() == 1) => {
+        "Iterator" | "IntoIterator" if matches!(actual, Type::Named { name, arguments } if name == "SequenceIterator" && arguments.len() == 1) => {
             true
         }
         "Iterator" | "IntoIterator" if matches!(actual, Type::Named { name, arguments } if name == "Range" && arguments.is_empty()) => {